refactor(course-table): extract ICS export logic to separate file
Move all ICS calendar export-related code from index.ts to a dedicated ics.ts module, and import the shared functions in the main course table feature file to reduce code duplication and improve maintainability.
